refactor(acp,tui): move session orchestration into a harness runtime (slice H2)
Move everything about launching and driving an ACP session out of the TUI
and into a new `nori_acp::runtime` module, per docs/specs/crate-layering.md
slice H ("TUI consumes harness events only"):

- `SessionLaunchSpec` — the frontend's inputs (agent, cwd, policies, MCP
  servers, version, session/fork context, optional `SessionResume`)
- `launch_session(spec) -> LaunchedSession { op_tx, handle, events }` —
  owns NoriConfig loading and AcpBackendConfig assembly (hooks,
  notifications, auto-worktree derivation, history, proxy), the
  connect/shutdown/timeout race (8s warning, 30s abort), the op-forwarding
  loop, and the session-control command loop
- `SessionEvent { Backend(Box<BackendEvent>), SpawnFailed, ShutdownRequested }`
- `AcpAgentCommand`, `AcpAgentHandle`, `drain_until_shutdown` move here
  from the TUI

tui/src/chatwidget/agent.rs shrinks 602 → 170 lines: registry precheck,
AgentConnecting emission, spec construction from the codex `Config`, and a
SessionEvent → AppEvent mapping loop. The two near-identical ~250-line
spawn/resume orchestration functions collapse into the single runtime path
(spawn-vs-resume error labels preserved).

Any frontend (headless exec mode, alternative UIs) can now launch sessions
through the harness without reimplementing orchestration.

Validation: full workspace suite green; tui-pty-e2e green (23 suites);
elizacp close-the-loop TUI drive green; just fmt + just fix clean.

Part of the crate-layering refactor (docs/specs/crate-layering.md).
