+++ nori-rs/docs.md
@@ -11,7 +11,7 @@ This is the Rust implementation of Nori, a terminal-based AI coding assistant. T
 The `nori-rs` directory is the root of a Cargo workspace containing all Rust code for the project. The workspace is organized into focused crates that handle specific concerns:
 
 - **Entry points**: `tui/` provides the main TUI application, `cli/` provides the `nori` binary dispatch and sandbox debug utilities
-- **ACP integration**: `acp/` (`nori-acp`) is the session harness -- it owns the ACP backend session runtime, transcripts, hooks, goals, and moved-in leaf helpers (shell parsing, notifications, custom prompts, compact constants). The agent-agnostic hosting machinery (subprocess spawning, wire client, registry, translator) lives in `acp-host/` (`nori-acp-host`) and is re-exported through `nori-acp` so consumer paths are unchanged. The Nori config layer lives in `nori-config/` and is imported directly by the frontends (`nori-acp` uses it internally but no longer re-exports it)
+- **ACP integration**: `acp/` (`nori-acp`) is the session harness -- it owns the ACP backend session runtime, the frontend-facing session launch runtime (`acp/src/runtime.rs`: frontends build a `SessionLaunchSpec`, call `launch_session`, and consume `SessionEvent`s), transcripts, hooks, goals, and moved-in leaf helpers (shell parsing, notifications, custom prompts, compact constants). The agent-agnostic hosting machinery (subprocess spawning, wire client, registry, translator) lives in `acp-host/` (`nori-acp-host`) and is re-exported through `nori-acp` so consumer paths are unchanged. The Nori config layer lives in `nori-config/` and is imported directly by the frontends (`nori-acp` uses it internally but no longer re-exports it)
 - **Shared infrastructure**: `core/` contains configuration, authentication, and model/provider metadata consumed by the frontends (not by `acp/`)
 - **Protocol definitions**: `protocol/`, `app-server-protocol/`, `mcp-types/` define shared type vocabularies
 - **Sandboxing**: `sandbox/` (`codex-sandbox`) owns the sandboxed exec engine and platform sandbox selection; `linux-sandbox/`, `windows-sandbox-rs/`, `execpolicy/` provide the platform-specific pieces
@@ -19,7 +19,7 @@ The `nori-rs` directory is the root of a Cargo workspace containing all Rust cod
 
 Most shared crates still follow the inherited `codex-` prefix convention (for example `codex-core` and `codex-protocol`), while Nori-owned entrypoint crates now use `nori-` names such as `nori-acp`, `nori-protocol`, `nori-installed`, and `nori-tui`.
 
-The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), `nori-acp`'s dependency on `codex-core` fully severed, the sandboxed-execution engine extracted from `codex-core` into `codex-sandbox` (`@/nori-rs/sandbox/`), the config layer extracted into `nori-config` (`@/nori-rs/nori-config/`) with the frontends (`nori-tui`, `nori-cli`) rewired to import it directly (nori-acp's config re-exports are gone), and the agent-agnostic ACP hosting machinery extracted into the Layer-0 `nori-acp-host` crate (`@/nori-rs/acp-host/`).
+The workspace is converging on a three-layer structure -- publishable ACP-host leaves at the bottom, a headless session harness in the middle, thin frontends on top -- per `@/docs/specs/crate-layering.md`. Milestones already landed: dead Codex-engine subsystems deleted from `codex-core`, the `nori-tui` `nori-config` cargo feature removed (Nori config is the only path), protocol types imported directly from `codex-protocol` (core's re-export detour deleted), `nori-acp`'s dependency on `codex-core` fully severed, the sandboxed-execution engine extracted from `codex-core` into `codex-sandbox` (`@/nori-rs/sandbox/`), the config layer extracted into `nori-config` (`@/nori-rs/nori-config/`) with the frontends (`nori-tui`, `nori-cli`) rewired to import it directly (nori-acp's config re-exports are gone), the agent-agnostic ACP hosting machinery extracted into the Layer-0 `nori-acp-host` crate (`@/nori-rs/acp-host/`), and session spawn/resume orchestration moved out of the TUI's `chatwidget/agent.rs` into the harness session runtime (`@/nori-rs/acp/src/runtime.rs`), leaving the TUI as a thin adapter that maps `SessionEvent`s onto app events.
 
 ### Core Implementation
