+++ nori-rs/tui/src/chatwidget/agent.rs
@@ -1,137 +1,29 @@
-use std::sync::Arc;
-use std::time::Duration;
+//! Thin adapter between the harness session runtime and the TUI event loop.
+//!
+//! All session orchestration (backend config assembly, connect/shutdown/
+//! timeout race, op forwarding, session-control commands) lives in
+//! `nori_acp::runtime`; this module only builds a launch spec from the codex
+//! `Config` and maps [`SessionEvent`]s onto [`AppEvent`]s.
 
 use codex_core::config::Config;
 use codex_protocol::protocol::Op;
-use nori_acp::AcpBackend;
-use nori_acp::AcpBackendConfig;
-use nori_acp::AcpSessionSummary;
-use nori_acp::SessionConfigOption;
-use nori_acp::get_agent_config;
 use nori_acp::get_agent_display_name;
 use nori_acp::list_available_agents;
-use nori_config::HistoryPersistence;
-use nori_config::find_nori_home;
-use tokio::sync::mpsc;
-use tokio::sync::mpsc::UnboundedReceiver;
+use nori_acp::runtime::SessionEvent;
+use nori_acp::runtime::SessionLaunchSpec;
+use nori_acp::runtime::SessionResume;
+use nori_acp::runtime::launch_session;
 use tokio::sync::mpsc::UnboundedSender;
 use tokio::sync::mpsc::unbounded_channel;
-use tokio::sync::oneshot;
+
+#[cfg(test)]
+pub(crate) use nori_acp::runtime::AcpAgentCommand;
+pub(crate) use nori_acp::runtime::AcpAgentHandle;
+#[cfg(test)]
+pub(crate) use nori_acp::runtime::drain_until_shutdown;
 
 use crate::app_event::AppEvent;
 use crate::app_event_sender::AppEventSender;
-
-/// Duration before showing a warning that connection is taking too long.
-const CONNECT_WARNING_SECS: u64 = 8;
-/// Duration after the warning before forcibly aborting the connection attempt.
-const CONNECT_ABORT_SECS: u64 = 30;
-
-/// Drain ops from the channel, discarding everything except `Op::Shutdown`.
-/// Returns when `Op::Shutdown` is received or the channel is closed.
-pub(crate) async fn drain_until_shutdown(rx: &mut UnboundedReceiver<Op>) {
-    while let Some(op) = rx.recv().await {
-        if matches!(op, Op::Shutdown) {
-            return;
-        }
-    }
-}
-
-/// Two-phase timeout: warn after `CONNECT_WARNING_SECS`, abort after an
-/// additional `CONNECT_ABORT_SECS`.
-async fn spawn_timeout_sequence(app_event_tx: &AppEventSender) {
-    tokio::time::sleep(Duration::from_secs(CONNECT_WARNING_SECS)).await;
-    app_event_tx.send(AppEvent::CodexEvent(codex_protocol::protocol::Event {
-        id: String::new(),
-        msg: codex_protocol::protocol::EventMsg::Warning(codex_protocol::protocol::WarningEvent {
-            message: format!(
-                "Connection is taking longer than expected. \
-                 Will abort in {CONNECT_ABORT_SECS}s if still unresponsive."
-            ),
-        }),
-    }));
-    tokio::time::sleep(Duration::from_secs(CONNECT_ABORT_SECS)).await;
-}
-
-/// Command for controlling ACP session state exposed by the agent.
-pub(crate) enum AcpAgentCommand {
-    /// Get the current ACP session config snapshot.
-    GetSessionConfig {
-        response_tx: oneshot::Sender<Vec<SessionConfigOption>>,
-    },
-    /// Set an ACP session config option.
-    SetSessionConfigOption {
-        config_id: String,
-        value: String,
-        response_tx: oneshot::Sender<anyhow::Result<Vec<SessionConfigOption>>>,
-    },
-    /// List the agent's known sessions via ACP `session/list`.
-    ListSessions {
-        cwd: std::path::PathBuf,
-        response_tx: oneshot::Sender<anyhow::Result<Vec<AcpSessionSummary>>>,
-    },
-}
-
-/// Handle for communicating with an ACP agent.
-///
-/// This handle provides access to ACP session control operations in addition
-/// to the standard Op channel.
-#[derive(Clone)]
-pub(crate) struct AcpAgentHandle {
-    command_tx: mpsc::UnboundedSender<AcpAgentCommand>,
-}
-
-impl AcpAgentHandle {
-    #[cfg(test)]
-    pub(crate) fn from_command_tx(command_tx: mpsc::UnboundedSender<AcpAgentCommand>) -> Self {
-        Self { command_tx }
-    }
-
-    /// Get the current ACP session config snapshot from the agent.
-    pub async fn get_session_config(&self) -> Option<Vec<SessionConfigOption>> {
-        let (response_tx, response_rx) = oneshot::channel();
-        if self
-            .command_tx
-            .send(AcpAgentCommand::GetSessionConfig { response_tx })
-            .is_err()
-        {
-            return None;
-        }
-        response_rx.await.ok()
-    }
-
-    /// Set an ACP session config option value.
-    pub async fn set_session_config_option(
-        &self,
-        config_id: String,
-        value: String,
-    ) -> anyhow::Result<Vec<SessionConfigOption>> {
-        let (response_tx, response_rx) = oneshot::channel();
-        self.command_tx
-            .send(AcpAgentCommand::SetSessionConfigOption {
-                config_id,
-                value,
-                response_tx,
-            })
-            .map_err(|_| anyhow::anyhow!("ACP agent command channel closed"))?;
-        response_rx
-            .await
-            .map_err(|_| anyhow::anyhow!("ACP agent did not respond"))?
-    }
-
-    /// List the agent's known sessions via ACP `session/list`.
-    pub async fn list_sessions(
-        &self,
-        cwd: std::path::PathBuf,
-    ) -> anyhow::Result<Vec<AcpSessionSummary>> {
-        let (response_tx, response_rx) = oneshot::channel();
-        self.command_tx
-            .send(AcpAgentCommand::ListSessions { cwd, response_tx })
-            .map_err(|_| anyhow::anyhow!("ACP agent command channel closed"))?;
-        response_rx
-            .await
-            .map_err(|_| anyhow::anyhow!("ACP agent did not respond"))?
-    }
-}
 
 /// Result of spawning an agent, which may include an ACP handle for model control.
 pub(crate) struct SpawnAgentResult {
@@ -151,8 +43,8 @@ pub(crate) fn spawn_agent(
     app_event_tx: AppEventSender,
     fork_context: Option<String>,
 ) -> SpawnAgentResult {
-    match get_agent_config(&config.model) {
-        Ok(_) => spawn_acp_agent(config, app_event_tx, fork_context),
+    match nori_acp::get_agent_config(&config.model) {
+        Ok(_) => launch_acp_agent(config, app_event_tx, fork_context, None),
         Err(_) => {
             let agent_name = config.model;
             let known: Vec<String> = list_available_agents()
@@ -171,6 +63,27 @@ pub(crate) fn spawn_agent(
             }
         }
     }
+}
+
+/// Spawn an ACP agent backend that resumes a previous session.
+///
+/// If the agent supports `session/load`, server-side resume is used.
+/// Otherwise, falls back to client-side replay using the provided transcript.
+pub(crate) fn spawn_acp_agent_resume(
+    config: Config,
+    acp_session_id: Option<String>,
+    transcript: Option<nori_acp::transcript::Transcript>,
+    app_event_tx: AppEventSender,
+) -> SpawnAgentResult {
+    launch_acp_agent(
+        config,
+        app_event_tx,
+        None,
+        Some(SessionResume {
+            acp_session_id,
+            transcript,
+        }),
+    )
 }
 
 /// Spawn an agent that emits an error and opens the agent picker.
@@ -195,408 +108,60 @@ fn spawn_error_agent(
     codex_op_tx
 }
 
-/// Spawn an ACP agent backend.
-///
-/// This uses the `nori_acp` crate to spawn an agent subprocess and handle
-/// communication via the Agent Client Protocol.
-fn spawn_acp_agent(
+/// Launch a session via the harness runtime and forward its events into the
+/// TUI event loop.
+fn launch_acp_agent(
     config: Config,
     app_event_tx: AppEventSender,
     fork_context: Option<String>,
+    resume: Option<SessionResume>,
 ) -> SpawnAgentResult {
-    let (codex_op_tx, mut codex_op_rx) = unbounded_channel::<Op>();
-
-    // Create the ACP command channel for model and session-config operations.
-    let (agent_cmd_tx, mut agent_cmd_rx) = unbounded_channel::<AcpAgentCommand>();
-
-    let acp_handle = Some(AcpAgentHandle {
-        command_tx: agent_cmd_tx,
-    });
-
     // Emit "Connecting" status before spawning the backend
     let display_name = get_agent_display_name(&config.model);
     app_event_tx.send(AppEvent::AgentConnecting { display_name });
 
+    let spec = SessionLaunchSpec {
+        agent: config.model.clone(),
+        cwd: config.cwd.clone(),
+        approval_policy: config.approval_policy,
+        sandbox_policy: config.sandbox_policy.clone(),
+        notify: config.notify.clone(),
+        mcp_servers: config.mcp_servers.clone(),
+        mcp_oauth_credentials_store_mode: config.mcp_oauth_credentials_store_mode,
+        cli_version: env!("CARGO_PKG_VERSION").to_string(),
+        session_context: Some(include_str!("../../session_context.md").to_string()),
+        initial_context: fork_context,
+        resume,
+    };
+    let agent_name = config.model;
+
+    let mut session = launch_session(spec);
+    let acp_handle = Some(session.handle.clone());
+    let op_tx = session.op_tx.clone();
+
     tokio::spawn(async move {
-        // Create a single ACP backend → TUI channel for both control-plane
-        // and normalized session-domain events.
-        let (backend_event_tx, mut backend_event_rx) = mpsc::channel(32);
-
-        // Create ACP backend config from codex config
-        let nori_home = find_nori_home().unwrap_or_else(|_| config.cwd.clone());
-        // Load NoriConfig for ACP-specific settings (os_notifications)
-        let nori_config = nori_config::NoriConfig::load().unwrap_or_default();
-        // Detect auto-worktree repo root from the cwd path.
-        // When auto_worktree is enabled, cwd is {repo_root}/.worktrees/{name},
-        // so we can derive repo_root by going up two directories.
-        let auto_worktree_repo_root = if nori_config.auto_worktree.is_enabled() {
-            config
-                .cwd
-                .parent()
-                .filter(|p| p.file_name().is_some_and(|n| n == ".worktrees"))
-                .and_then(|p| p.parent())
-                .map(std::path::Path::to_path_buf)
-        } else {
-            None
-        };
-        // Resolve to Off if no worktree actually exists (e.g. "ask" mode
-        // where the user declined).
-        let auto_worktree = if auto_worktree_repo_root.is_some() {
-            nori_config.auto_worktree
-        } else {
-            nori_config::AutoWorktree::Off
-        };
-
-        let acp_config = AcpBackendConfig {
-            agent: config.model.clone(),
-            cwd: config.cwd.clone(),
-            approval_policy: config.approval_policy,
-            sandbox_policy: config.sandbox_policy.clone(),
-            notify: config.notify.clone(),
-            os_notifications: nori_config.os_notifications,
-            notify_after_idle: nori_config.notify_after_idle,
-            nori_home,
-            history_persistence: HistoryPersistence::SaveAll,
-            acp_proxy: nori_config.acp_proxy.clone(),
-            cli_version: env!("CARGO_PKG_VERSION").to_string(),
-            auto_worktree,
-            auto_worktree_repo_root,
-            session_start_hooks: nori_config.session_start_hooks.clone(),
-            session_end_hooks: nori_config.session_end_hooks.clone(),
-            pre_user_prompt_hooks: nori_config.pre_user_prompt_hooks.clone(),
-            post_user_prompt_hooks: nori_config.post_user_prompt_hooks.clone(),
-            pre_tool_call_hooks: nori_config.pre_tool_call_hooks.clone(),
-            post_tool_call_hooks: nori_config.post_tool_call_hooks.clone(),
-            pre_agent_response_hooks: nori_config.pre_agent_response_hooks.clone(),
-            post_agent_response_hooks: nori_config.post_agent_response_hooks.clone(),
-            async_session_start_hooks: nori_config.async_session_start_hooks.clone(),
-            async_session_end_hooks: nori_config.async_session_end_hooks.clone(),
-            async_pre_user_prompt_hooks: nori_config.async_pre_user_prompt_hooks.clone(),
-            async_post_user_prompt_hooks: nori_config.async_post_user_prompt_hooks.clone(),
-            async_pre_tool_call_hooks: nori_config.async_pre_tool_call_hooks.clone(),
-            async_post_tool_call_hooks: nori_config.async_post_tool_call_hooks.clone(),
-            async_pre_agent_response_hooks: nori_config.async_pre_agent_response_hooks.clone(),
-            async_post_agent_response_hooks: nori_config.async_post_agent_response_hooks.clone(),
-            script_timeout: nori_config.script_timeout.as_duration(),
-            default_model: nori_config.default_models.get(&config.model).cloned(),
-            initial_context: fork_context,
-            session_context: Some(include_str!("../../session_context.md").to_string()),
-            mcp_servers: config.mcp_servers.clone(),
-            mcp_oauth_credentials_store_mode: config.mcp_oauth_credentials_store_mode,
-        };
-
-        // Race backend init against shutdown requests and a timeout.
-        // This ensures the user can always exit even if the backend hangs.
-        let backend = tokio::select! {
-            result = AcpBackend::spawn(&acp_config, backend_event_tx) => {
-                match result {
-                    Ok(b) => Arc::new(b),
-                    Err(e) => {
-                        tracing::error!("failed to spawn ACP backend: {e}");
-                        drop(codex_op_rx);
-                        app_event_tx.send(AppEvent::AgentSpawnFailed {
-                            agent_name: config.model.clone(),
-                            error: format!("Failed to spawn ACP agent: {e}"),
-                        });
-                        return;
-                    }
-                }
-            }
-            () = drain_until_shutdown(&mut codex_op_rx) => {
-                tracing::info!("shutdown requested while ACP backend was connecting");
-                drop(codex_op_rx);
-                app_event_tx.send(AppEvent::ExitRequest);
-                return;
-            }
-            () = spawn_timeout_sequence(&app_event_tx) => {
-                tracing::warn!("ACP backend connection timed out");
-                drop(codex_op_rx);
-                app_event_tx.send(AppEvent::AgentSpawnFailed {
-                    agent_name: config.model.clone(),
-                    error: "Connection timed out. The agent did not respond.".to_string(),
-                });
-                return;
-            }
-        };
-
-        // Forward ops to backend
-        let backend_for_ops = Arc::clone(&backend);
-        tokio::spawn(async move {
-            while let Some(op) = codex_op_rx.recv().await {
-                if let Err(e) = backend_for_ops.submit(op).await {
-                    tracing::error!("failed to submit op: {e}");
-                }
-            }
-        });
-
-        let backend_for_agent = Arc::clone(&backend);
-        tokio::spawn(async move {
-            while let Some(cmd) = agent_cmd_rx.recv().await {
-                match cmd {
-                    AcpAgentCommand::GetSessionConfig { response_tx } => {
-                        let state = backend_for_agent.config_options();
-                        let _ = response_tx.send(state);
-                    }
-                    AcpAgentCommand::SetSessionConfigOption {
-                        config_id,
-                        value,
-                        response_tx,
-                    } => {
-                        let result = backend_for_agent
-                            .set_config_option(config_id, value)
-                            .await
-                            .map(|()| backend_for_agent.config_options());
-                        let _ = response_tx.send(result);
-                    }
-                    AcpAgentCommand::ListSessions { cwd, response_tx } => {
-                        let result = backend_for_agent.connection().list_sessions(&cwd).await;
-                        let _ = response_tx.send(result);
-                    }
-                }
-            }
-        });
-
-        // Drop our Arc reference - the op and agent-control tasks have their own.
-        // This is necessary so that when these tasks exit, the backend is fully dropped,
-        // which drops event_tx, allowing event_rx to return None and this task to exit.
-        drop(backend);
-
-        while let Some(event) = backend_event_rx.recv().await {
+        while let Some(event) = session.events.recv().await {
             match event {
-                nori_acp::BackendEvent::Control(event) => {
-                    app_event_tx.send(AppEvent::CodexEvent(event));
+                SessionEvent::Backend(backend_event) => match *backend_event {
+                    nori_acp::BackendEvent::Control(event) => {
+                        app_event_tx.send(AppEvent::CodexEvent(event));
+                    }
+                    nori_acp::BackendEvent::Client(client_event) => {
+                        app_event_tx.send(AppEvent::ClientEvent(client_event));
+                    }
+                },
+                SessionEvent::SpawnFailed { error } => {
+                    app_event_tx.send(AppEvent::AgentSpawnFailed {
+                        agent_name: agent_name.clone(),
+                        error,
+                    });
                 }
-                nori_acp::BackendEvent::Client(client_event) => {
-                    app_event_tx.send(AppEvent::ClientEvent(client_event));
+                SessionEvent::ShutdownRequested => {
+                    app_event_tx.send(AppEvent::ExitRequest);
                 }
             }
         }
     });
 
-    SpawnAgentResult {
-        op_tx: codex_op_tx,
-        acp_handle,
-    }
-}
-
-/// Spawn an ACP agent backend that resumes a previous session.
-///
-/// Similar to `spawn_acp_agent`, but calls `AcpBackend::resume_session`
-/// instead of `AcpBackend::spawn`. If the agent supports `session/load`,
-/// server-side resume is used. Otherwise, falls back to client-side replay
-/// using the provided transcript.
-pub(crate) fn spawn_acp_agent_resume(
-    config: Config,
-    acp_session_id: Option<String>,
-    transcript: Option<nori_acp::transcript::Transcript>,
-    app_event_tx: AppEventSender,
-) -> SpawnAgentResult {
-    let (codex_op_tx, mut codex_op_rx) = unbounded_channel::<Op>();
-
-    let (agent_cmd_tx, mut agent_cmd_rx) = unbounded_channel::<AcpAgentCommand>();
-
-    let acp_handle = Some(AcpAgentHandle {
-        command_tx: agent_cmd_tx,
-    });
-
-    let display_name = get_agent_display_name(&config.model);
-    app_event_tx.send(AppEvent::AgentConnecting { display_name });
-
-    tokio::spawn(async move {
-        let (backend_event_tx, mut backend_event_rx) = mpsc::channel(32);
-
-        let nori_home = find_nori_home().unwrap_or_else(|_| config.cwd.clone());
-        let nori_config = nori_config::NoriConfig::load().unwrap_or_default();
-        let auto_worktree_repo_root = if nori_config.auto_worktree.is_enabled() {
-            config
-                .cwd
-                .parent()
-                .filter(|p| p.file_name().is_some_and(|n| n == ".worktrees"))
-                .and_then(|p| p.parent())
-                .map(std::path::Path::to_path_buf)
-        } else {
-            None
-        };
-        // Resolve to Off if no worktree actually exists (e.g. "ask" mode
-        // where the user declined).
-        let auto_worktree = if auto_worktree_repo_root.is_some() {
-            nori_config.auto_worktree
-        } else {
-            nori_config::AutoWorktree::Off
-        };
-
-        let acp_config = AcpBackendConfig {
-            agent: config.model.clone(),
-            cwd: config.cwd.clone(),
-            approval_policy: config.approval_policy,
-            sandbox_policy: config.sandbox_policy.clone(),
-            notify: config.notify.clone(),
-            os_notifications: nori_config.os_notifications,
-            notify_after_idle: nori_config.notify_after_idle,
-            nori_home,
-            history_persistence: HistoryPersistence::SaveAll,
-            acp_proxy: nori_config.acp_proxy.clone(),
-            cli_version: env!("CARGO_PKG_VERSION").to_string(),
-            auto_worktree,
-            auto_worktree_repo_root,
-            session_start_hooks: nori_config.session_start_hooks.clone(),
-            session_end_hooks: nori_config.session_end_hooks.clone(),
-            pre_user_prompt_hooks: nori_config.pre_user_prompt_hooks.clone(),
-            post_user_prompt_hooks: nori_config.post_user_prompt_hooks.clone(),
-            pre_tool_call_hooks: nori_config.pre_tool_call_hooks.clone(),
-            post_tool_call_hooks: nori_config.post_tool_call_hooks.clone(),
-            pre_agent_response_hooks: nori_config.pre_agent_response_hooks.clone(),
-            post_agent_response_hooks: nori_config.post_agent_response_hooks.clone(),
-            async_session_start_hooks: nori_config.async_session_start_hooks.clone(),
-            async_session_end_hooks: nori_config.async_session_end_hooks.clone(),
-            async_pre_user_prompt_hooks: nori_config.async_pre_user_prompt_hooks.clone(),
-            async_post_user_prompt_hooks: nori_config.async_post_user_prompt_hooks.clone(),
-            async_pre_tool_call_hooks: nori_config.async_pre_tool_call_hooks.clone(),
-            async_post_tool_call_hooks: nori_config.async_post_tool_call_hooks.clone(),
-            async_pre_agent_response_hooks: nori_config.async_pre_agent_response_hooks.clone(),
-            async_post_agent_response_hooks: nori_config.async_post_agent_response_hooks.clone(),
-            script_timeout: nori_config.script_timeout.as_duration(),
-            default_model: nori_config.default_models.get(&config.model).cloned(),
-            initial_context: None,
-            session_context: Some(include_str!("../../session_context.md").to_string()),
-            mcp_servers: config.mcp_servers.clone(),
-            mcp_oauth_credentials_store_mode: config.mcp_oauth_credentials_store_mode,
-        };
-
-        // Race backend resume against shutdown requests and a timeout.
-        let backend = tokio::select! {
-            result = AcpBackend::resume_session(
-                &acp_config,
-                acp_session_id.as_deref(),
-                transcript.as_ref(),
-                backend_event_tx,
-            ) => {
-                match result {
-                    Ok(b) => Arc::new(b),
-                    Err(e) => {
-                        tracing::error!("failed to resume ACP session: {e}");
-                        drop(codex_op_rx);
-                        app_event_tx.send(AppEvent::AgentSpawnFailed {
-                            agent_name: config.model.clone(),
-                            error: format!("Failed to resume ACP session: {e}"),
-                        });
-                        return;
-                    }
-                }
-            }
-            () = drain_until_shutdown(&mut codex_op_rx) => {
-                tracing::info!("shutdown requested while resuming ACP session");
-                drop(codex_op_rx);
-                app_event_tx.send(AppEvent::ExitRequest);
-                return;
-            }
-            () = spawn_timeout_sequence(&app_event_tx) => {
-                tracing::warn!("ACP session resume timed out");
-                drop(codex_op_rx);
-                app_event_tx.send(AppEvent::AgentSpawnFailed {
-                    agent_name: config.model.clone(),
-                    error: "Connection timed out. The agent did not respond.".to_string(),
-                });
-                return;
-            }
-        };
-
-        let backend_for_ops = Arc::clone(&backend);
-        tokio::spawn(async move {
-            while let Some(op) = codex_op_rx.recv().await {
-                if let Err(e) = backend_for_ops.submit(op).await {
-                    tracing::error!("failed to submit op: {e}");
-                }
-            }
-        });
-
-        let backend_for_agent = Arc::clone(&backend);
-        tokio::spawn(async move {
-            while let Some(cmd) = agent_cmd_rx.recv().await {
-                match cmd {
-                    AcpAgentCommand::GetSessionConfig { response_tx } => {
-                        let state = backend_for_agent.config_options();
-                        let _ = response_tx.send(state);
-                    }
-                    AcpAgentCommand::SetSessionConfigOption {
-                        config_id,
-                        value,
-                        response_tx,
-                    } => {
-                        let result = backend_for_agent
-                            .set_config_option(config_id, value)
-                            .await
-                            .map(|()| backend_for_agent.config_options());
-                        let _ = response_tx.send(result);
-                    }
-                    AcpAgentCommand::ListSessions { cwd, response_tx } => {
-                        let result = backend_for_agent.connection().list_sessions(&cwd).await;
-                        let _ = response_tx.send(result);
-                    }
-                }
-            }
-        });
-
-        drop(backend);
-
-        while let Some(event) = backend_event_rx.recv().await {
-            match event {
-                nori_acp::BackendEvent::Control(event) => {
-                    app_event_tx.send(AppEvent::CodexEvent(event));
-                }
-                nori_acp::BackendEvent::Client(client_event) => {
-                    app_event_tx.send(AppEvent::ClientEvent(client_event));
-                }
-            }
-        }
-    });
-
-    SpawnAgentResult {
-        op_tx: codex_op_tx,
-        acp_handle,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    fn mode_option(current_value: &str) -> SessionConfigOption {
-        SessionConfigOption::select(
-            "mode",
-            "Mode",
-            current_value.to_string(),
-            vec![
-                nori_acp::SessionConfigSelectOption::new("plan", "Plan"),
-                nori_acp::SessionConfigSelectOption::new("build", "Build"),
-            ],
-        )
-        .category(nori_acp::SessionConfigOptionCategory::Mode)
-    }
-
-    #[tokio::test]
-    async fn set_session_config_option_returns_refreshed_config_snapshot() {
-        let (command_tx, mut command_rx) = unbounded_channel::<AcpAgentCommand>();
-        tokio::spawn(async move {
-            while let Some(command) = command_rx.recv().await {
-                if let AcpAgentCommand::SetSessionConfigOption {
-                    config_id: _,
-                    value,
-                    response_tx,
-                } = command
-                {
-                    let _ = response_tx.send(Ok(vec![mode_option(&value)]));
-                }
-            }
-        });
-        let handle = AcpAgentHandle { command_tx };
-
-        let config_options = handle
-            .set_session_config_option("mode".to_string(), "build".to_string())
-            .await
-            .unwrap();
-
-        assert_eq!(config_options, vec![mode_option("build")]);
-    }
+    SpawnAgentResult { op_tx, acp_handle }
 }
